requirements.txt
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keras
Markdown==3.0.1
numpy==1.15.0
protobuf==3.7.0
six==1.12.0
tensorflow==1.13.1
tensorflow-probability==0.6.0
tfmpl==1.0.2
termcolor==1.1.0
Werkzeug==0.15.1
